% Auto-generated: do not edit by hand
\name{dashImageGallery}

\alias{dashImageGallery}

\title{DashImageGallery component}

\description{

}

\usage{
dashImageGallery(id=NULL, additionalClass=NULL, autoPlay=NULL,
disableKeyDown=NULL, disableSwipe=NULL,
disableThumbnailScroll=NULL, disableThumbnailSwipe=NULL,
flickThreshold=NULL, indexSeparator=NULL, infinite=NULL,
isRTL=NULL, items=NULL, lazyLoad=NULL, onBeforeSlide=NULL,
onBulletClick=NULL, onClick=NULL, onErrorImageURL=NULL,
onImageError=NULL, onImageLoad=NULL, onMouseLeave=NULL,
onMouseOver=NULL, onPause=NULL, onPlay=NULL,
onScreenChange=NULL, onSlide=NULL, onThumbnailClick=NULL,
onThumbnailError=NULL, onTouchEnd=NULL, onTouchMove=NULL,
onTouchStart=NULL, renderCustomControls=NULL,
renderFullscreenButton=NULL, renderItem=NULL,
renderLeftNav=NULL, renderPlayPauseButton=NULL,
renderRightNav=NULL, renderThumbInner=NULL,
showBullets=NULL, showFullscreenButton=NULL, showIndex=NULL,
showNav=NULL, showPlayButton=NULL, showThumbnails=NULL,
slideDuration=NULL, slideInterval=NULL,
slideOnThumbnailOver=NULL, startIndex=NULL,
stopPropagation=NULL, swipeThreshold=NULL,
swipingTransitionDuration=NULL, thumbnailPosition=NULL,
useBrowserFullscreen=NULL, useTranslate3D=NULL,
useWindowKeyDown=NULL)
}

\arguments{
\item{id}{Character. }

\item{additionalClass}{Character. }

\item{autoPlay}{Logical. }

\item{disableKeyDown}{Logical. }

\item{disableSwipe}{Logical. }

\item{disableThumbnailScroll}{Logical. }

\item{disableThumbnailSwipe}{Logical. }

\item{flickThreshold}{Numeric. }

\item{indexSeparator}{Character. }

\item{infinite}{Logical. }

\item{isRTL}{Logical. }

\item{items}{List of lists containing elements 'original', 'thumbnail', 'fullscreen', 'originalheight', 'originalwidth', 'loading', 'thumbnailheight', 'thumbnailwidth', 'thumbnailloading', 'originalclass', 'thumbnailclass', 'renderitem', 'renderthumbinner', 'originalalt', 'thumbnailalt', 'originaltitle', 'thumbnailtitle', 'thumbnaillabel', 'description', 'srcset', 'sizes', 'bulletclass'.
those elements have the following types:
  - original (character; optional)
  - thumbnail (character; optional)
  - fullscreen (character; optional)
  - originalheight (numeric; optional)
  - originalwidth (numeric; optional)
  - loading (character; optional)
  - thumbnailheight (numeric; optional)
  - thumbnailwidth (numeric; optional)
  - thumbnailloading (character; optional)
  - originalclass (character; optional)
  - thumbnailclass (character; optional)
  - renderitem (optional)
  - renderthumbinner (optional)
  - originalalt (character; optional)
  - thumbnailalt (character; optional)
  - originaltitle (character; optional)
  - thumbnailtitle (character; optional)
  - thumbnaillabel (character; optional)
  - description (character; optional)
  - srcset (character; optional)
  - sizes (character; optional)
  - bulletclass (character; optional)s. }

\item{lazyLoad}{Logical. }

\item{onBeforeSlide}{}

\item{onBulletClick}{}

\item{onClick}{}

\item{onErrorImageURL}{Character. }

\item{onImageError}{}

\item{onImageLoad}{}

\item{onMouseLeave}{}

\item{onMouseOver}{}

\item{onPause}{}

\item{onPlay}{}

\item{onScreenChange}{}

\item{onSlide}{}

\item{onThumbnailClick}{}

\item{onThumbnailError}{}

\item{onTouchEnd}{}

\item{onTouchMove}{}

\item{onTouchStart}{}

\item{renderCustomControls}{}

\item{renderFullscreenButton}{}

\item{renderItem}{}

\item{renderLeftNav}{}

\item{renderPlayPauseButton}{}

\item{renderRightNav}{}

\item{renderThumbInner}{}

\item{showBullets}{Logical. }

\item{showFullscreenButton}{Logical. }

\item{showIndex}{Logical. }

\item{showNav}{Logical. }

\item{showPlayButton}{Logical. }

\item{showThumbnails}{Logical. }

\item{slideDuration}{Numeric. }

\item{slideInterval}{Numeric. }

\item{slideOnThumbnailOver}{Logical. }

\item{startIndex}{Numeric. }

\item{stopPropagation}{Logical. }

\item{swipeThreshold}{Numeric. }

\item{swipingTransitionDuration}{Numeric. }

\item{thumbnailPosition}{Character. }

\item{useBrowserFullscreen}{Logical. }

\item{useTranslate3D}{Logical. }

\item{useWindowKeyDown}{Logical. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

